broke everytying to get the static definition out of GPXFile Firstly, it's a bad call to have statics; global variables anyone. Secondly, it contained definitely non-static variable types which consequently had to be declared 'static' just to get the system to compile. It was broken.  Thirdly, when you kotow to this and change your variable to static, then you lose the ability to have derived classes - they are not checked at runtime in static variables. So, since the new line rendering code does lots of derived classes, it was fundamentally incompatible with the use of static. SO.  So, I changed the whole system to remove statics where they were really just baggage.
