+++ OsmAnd/src/net/osmand/plus/GPXUtilities.java
@@ -2,13 +2,22 @@
 package net.osmand.plus;
 
 import android.content.Context;
+import android.graphics.Canvas;
 import android.graphics.Color;
+import android.graphics.Paint;
+import android.graphics.Path;
+import android.os.AsyncTask;
 
 import net.osmand.Location;
 import net.osmand.PlatformUtil;
 import net.osmand.data.LocationPoint;
 import net.osmand.data.PointDescription;
+import net.osmand.data.QuadRect;
+import net.osmand.data.RotatedTileBox;
+import net.osmand.plus.views.OsmandMapTileView;
 import net.osmand.util.Algorithms;
+import net.osmand.util.MapAlgorithms;
+import net.osmand.util.MapUtils;
 
 import org.apache.commons.logging.Log;
 import org.xmlpull.v1.XmlPullParser;
@@ -34,493 +43,44 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 import java.util.ArrayList;
 import java.util.Collections;
+import java.util.Comparator;
 import java.util.Date;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Locale;
 import java.util.Map;
+import java.util.Set;
 import java.util.Stack;
 import java.util.TimeZone;
 
+import gnu.trove.list.array.TIntArrayList;
+
+
 public class GPXUtilities {
+
 	public final static Log log = PlatformUtil.getLog(GPXUtilities.class);
 
 	private final static String GPX_TIME_FORMAT = "yyyy-MM-dd'T'HH:mm:ss'Z'"; //$NON-NLS-1$
 
-	private final static NumberFormat latLonFormat = new DecimalFormat("0.00#####", new DecimalFormatSymbols(
-			new Locale("EN", "US")));
 
-	public static class GPXExtensions {
-		Map<String, String> extensions = null;
-
-		public Map<String, String> getExtensionsToRead() {
-			if (extensions == null) {
-				return Collections.emptyMap();
-			}
-			return extensions;
-		}
-
-		private boolean customZoom = false;
-		public boolean hasCustomZoom() {
-			customZoom = customZoom | (extensions != null && extensions.containsKey("zoom"));
-			return customZoom;
-		}
-		public void setCustomZoom(boolean zoom) {
-			customZoom = zoom;
-		}
-
-		public float getGpxZoom(float defaultGpxZoom) {
-			if(extensions != null && extensions.containsKey("zoom")) {
-				try {
-					defaultGpxZoom = Float.parseFloat(extensions.get("zoom"));
-				} catch (NumberFormatException e) {
-					e.printStackTrace();
-				}
-			}
-			return defaultGpxZoom;
-		}
-
-		public void setGpxZoom(float gpxZoom) {
-			getExtensionsToWrite().put("zoom", Float.toString(gpxZoom));
-		}
-
-		public int getColor(int defColor) {
-			if(extensions != null && extensions.containsKey("color")) {
-				try {
-					return Color.parseColor(extensions.get("color").toUpperCase());
-				} catch (IllegalArgumentException e) {
-					e.printStackTrace();
-				}
-			}
-			return defColor;
-		}
-		
-		public void setColor(int color) {
-			getExtensionsToWrite().put("color", Algorithms.colorToString(color));
-		}
-
-		public Map<String, String> getExtensionsToWrite() {
-			if (extensions == null) {
-				extensions = new LinkedHashMap<String, String>();
-			}
-			return extensions;
-		}
-
+	public List<GPXTrackAnalysis> splitByDistance(double meters) {
+		return split(getDistanceMetric(), getTimeSplit(), meters);
 	}
 
-	public static class WptPt extends GPXExtensions implements LocationPoint {
-		public double lat;
-		public double lon;
-		public String name = null;
-		public String link = null;
-		// previous undocumented feature 'category' ,now 'type'
-		public String category = null;
-		public String desc = null;
-		// by default
-		public long time = 0;
-		public double ele = Double.NaN;
-		public double speed = 0;
-		public double hdop = Double.NaN;
-		public boolean deleted = false;
-
-		public WptPt() {
-		}
-
-		@Override
-		public int getColor() {
-			return getColor(0);
-		}
-
-		@Override
-		public double getLatitude() {
-			return lat;
-		}
-
-		@Override
-		public double getLongitude() {
-			return lon;
-		}
-
-		
-		@Override
-		public PointDescription getPointDescription(Context ctx) {
-			return new PointDescription(PointDescription.POINT_TYPE_WPT, name);
-		}
-
-		public WptPt(double lat, double lon, long time, double ele, double speed, double hdop) {
-			this.lat = lat;
-			this.lon = lon;
-			this.time = time;
-			this.ele = ele;
-			this.speed = speed;
-			this.hdop = hdop;
-		}
-
-		@Override
-		public boolean isVisible() {
-			return true;
-		}
-
-		@Override
-		public int hashCode() {
-			final int prime = 31;
-			int result = 1;
-			result = prime * result + ((name == null) ? 0 : name.hashCode());
-			result = prime * result + ((category == null) ? 0 : category.hashCode());
-			result = prime * result + ((desc == null) ? 0 : desc.hashCode());
-			result = prime * result + ((lat == 0) ? 0 : Double.valueOf(lat).hashCode());
-			result = prime * result + ((lon == 0) ? 0 : Double.valueOf(lon).hashCode());
-			return result;
-		}
-
-		@Override
-		public boolean equals(Object obj) {
-			if (this == obj)
-				return true;
-			if (obj == null || getClass() != obj.getClass())
-				return false;
-			WptPt other = (WptPt) obj;
-			return Algorithms.objectEquals(other.name, name)
-					&& Algorithms.objectEquals(other.category, category)
-					&& Algorithms.objectEquals(other.lat, lat)
-					&& Algorithms.objectEquals(other.lon, lon)
-					&& Algorithms.objectEquals(other.desc, desc);
-		}
+	public List<GPXTrackAnalysis> splitByTime(int seconds) {
+		return split(getTimeSplit(), getDistanceMetric(), seconds);
 	}
 
-	public static class TrkSegment extends GPXExtensions {
-		public List<WptPt> points = new ArrayList<WptPt>();
-		public List<WptPt> culledPoints = null;
-		public int culledFingerprint = -1;
-
-		public List<GPXTrackAnalysis> splitByDistance(double meters) {
-			return split(getDistanceMetric(), getTimeSplit(), meters);
-		}
-		
-		public List<GPXTrackAnalysis> splitByTime(int seconds) {
-			return split(getTimeSplit(), getDistanceMetric(), seconds);
-		}
-		
-		private List<GPXTrackAnalysis> split(SplitMetric metric, SplitMetric secondaryMetric, double metricLimit) {
-			List<SplitSegment> splitSegments = new ArrayList<GPXUtilities.SplitSegment>();
-			splitSegment(metric, secondaryMetric, metricLimit, splitSegments, this);
-			return convert(splitSegments);
-		}
-
+	private List<GPXTrackAnalysis> split(SplitMetric metric, SplitMetric secondaryMetric, double metricLimit) {
+		List<SplitSegment> splitSegments = new ArrayList<SplitSegment>();
+	//TODO BOO????	splitSegment(metric, secondaryMetric, metricLimit, splitSegments, this);			//<<???
+		return convert(splitSegments);
 	}
 
-	public static class Track extends GPXExtensions {
-		public String name = null;
-		public String desc = null;
-		public List<TrkSegment> segments = new ArrayList<TrkSegment>();
-
-	}
-
-	public static class Route extends GPXExtensions {
-		public String name = null;
-		public String desc = null;
-		public List<WptPt> points = new ArrayList<WptPt>();
-
-	}
-	
-	public static class GPXTrackAnalysis {
-		public float totalDistance = 0;
-		public int totalTracks = 0;
-		public long startTime = Long.MAX_VALUE;
-		public long endTime = Long.MIN_VALUE;
-		public long timeSpan = 0;
-		public long timeMoving = 0;
-		public float totalDistanceMoving = 0;
-
-		public double diffElevationUp = 0;
-		public double diffElevationDown = 0;
-		public double avgElevation = 0;
-		public double minElevation = 99999;
-		public double maxElevation = -100;
-		
-		public float maxSpeed = 0;
-		public float avgSpeed;
-		
-		public int points;
-		public int wptPoints = 0;
-		
-		public double metricEnd;
-		public double secondaryMetricEnd;
-		public WptPt locationStart;
-		public WptPt locationEnd;
-
-		public boolean isTimeSpecified() {
-			return startTime != Long.MAX_VALUE && startTime != 0;
-		}
-		
-		public boolean isTimeMoving() {
-			return timeMoving != 0;
-		}
-		
-		public boolean isElevationSpecified() {
-			return maxElevation != -100;
-		}
-		
-		public boolean isSpeedSpecified() {
-			return avgSpeed > 0;
-		}
-		
-		
-		public static GPXTrackAnalysis segment(long filetimestamp, TrkSegment segment) {
-			return new GPXTrackAnalysis().prepareInformation(filetimestamp, new SplitSegment(segment));
-		}
-		
-		public GPXTrackAnalysis prepareInformation(long filestamp, SplitSegment... splitSegments) {
-			float[] calculations = new float[1];
-
-			float totalElevation = 0;
-			int elevationPoints = 0;
-			int speedCount = 0;
-			double totalSpeedSum = 0;
-			points = 0;
-			
-			double channelThresMin = 5;            // Minimum oscillation amplitude considered as noise for Up/Down analysis
-			double channelThres = channelThresMin; // Actual oscillation amplitude considered as noise, try depedency on current hdop
-			double channelBase;
-			double channelTop;
-			double channelBottom;
-			boolean climb = false;
-
-			for (SplitSegment s : splitSegments) {
-				final int numberOfPoints = s.getNumberOfPoints();
-
-				channelBase = 99999;
-				channelTop = channelBase;
-				channelBottom = channelBase;
-				channelThres = channelThresMin;
-
-				metricEnd += s.metricEnd;
-				secondaryMetricEnd += s.secondaryMetricEnd;
-				points += numberOfPoints;
-				for (int j = 0; j < numberOfPoints; j++) {
-					WptPt point = s.get(j);
-					if(j == 0 && locationStart == null) {
-						locationStart = point;
-					}
-					if(j == numberOfPoints - 1) {
-						locationEnd = point;
-					}
-					long time = point.time;
-					if (time != 0) {
-						startTime = Math.min(startTime, time);
-						endTime = Math.max(endTime, time);
-					}
-
-					double elevation = point.ele;
-					if (!Double.isNaN(elevation)) {
-						totalElevation += elevation;
-						elevationPoints++;
-						minElevation = Math.min(elevation, minElevation);
-						maxElevation = Math.max(elevation, maxElevation);
-					}
-
-					float speed = (float) point.speed;
-					if (speed > 0) {
-						totalSpeedSum += speed;
-						maxSpeed = Math.max(speed, maxSpeed);
-						speedCount++;
-					}
-
-					// Trend channel approach for elevation gain/loss, Hardy 2015-09-22
-					// Self-adjusting turnarund threshold added for testing 2015-09-25: Current rule is now: "All up/down trends of amplitude <X are ignored to smooth the noise, where X is the maximum observed DOP value of any point which contributed to the current trend (but at least 5 m as the minimum noise threshold)".
-					if (!Double.isNaN(point.ele)) {
-						// Init channel
-						if (channelBase == 99999) {
-							channelBase = point.ele;
-							channelTop = channelBase;
-							channelBottom = channelBase;
-							channelThres = channelThresMin;
-						}
-						// Channel maintenance
-						if (point.ele > channelTop) {
-							channelTop = point.ele;
-							if (!Double.isNaN(point.hdop)) {
-								channelThres = Math.max(channelThres, 2.0*point.hdop);  //Try empirical 2*hdop, may better serve very flat tracks, or high dop tracks
-							}
-						} else if (point.ele < channelBottom) {
-							channelBottom = point.ele;
-							if (!Double.isNaN(point.hdop)) {
-								channelThres = Math.max(channelThres, 2.0*point.hdop);
-							}
-						}
-						// Turnaround (breakout) detection
-						if ((point.ele <= (channelTop - channelThres)) && (climb == true)) {
-							if ((channelTop - channelBase) >= channelThres) {
-								diffElevationUp += channelTop - channelBase;
-							}
-							channelBase = channelTop;
-							channelBottom = point.ele;
-							climb = false;
-							channelThres = channelThresMin;
-						} else if ((point.ele >= (channelBottom + channelThres)) && (climb == false)) {
-							if ((channelBase - channelBottom) >= channelThres) {
-								diffElevationDown += channelBase - channelBottom;
-							}
-							channelBase = channelBottom;
-							channelTop = point.ele;
-							climb = true;
-							channelThres = channelThresMin;
-						}
-						// End detection without breakout
-						if (j == (numberOfPoints -1)) {
-							if ((channelTop - channelBase) >= channelThres) {
-								diffElevationUp += channelTop - channelBase;
-							}
-							if ((channelBase - channelBottom) >= channelThres) {
-								diffElevationDown += channelBase - channelBottom;
-							}
-						}
-					}
-
-					if (j > 0) {
-						WptPt prev = s.get(j - 1);
-
-						// Old complete summation approach for elevation gain/loss
-						//if (!Double.isNaN(point.ele) && !Double.isNaN(prev.ele)) {
-						//	double diff = point.ele - prev.ele;
-						//	if (diff > 0) {
-						//		diffElevationUp += diff;
-						//	} else {
-						//		diffElevationDown -= diff;
-						//	}
-						//}
-
-						// totalDistance += MapUtils.getDistance(prev.lat, prev.lon, point.lat, point.lon);
-						// using ellipsoidal 'distanceBetween' instead of spherical haversine (MapUtils.getDistance) is
-						// a little more exact, also seems slightly faster:
-						net.osmand.Location.distanceBetween(prev.lat, prev.lon, point.lat, point.lon, calculations);
-						totalDistance += calculations[0];
-
-						// Averaging speed values is less exact than totalDistance/timeMoving
-						if (speed > 0 && point.time != 0 && prev.time != 0) {
-							timeMoving = timeMoving + (point.time - prev.time);
-							totalDistanceMoving += calculations[0];
-						}
-					}
-				}
-			}
-			if(!isTimeSpecified()){
-				startTime = filestamp;
-				endTime = filestamp;
-			}
-
-			// OUTPUT:
-			// 1. Total distance, Start time, End time
-			// 2. Time span
-			timeSpan = endTime - startTime;
-
-			// 3. Time moving, if any
-			// 4. Elevation, eleUp, eleDown, if recorded
-			if (elevationPoints > 0) {
-				avgElevation =  totalElevation / elevationPoints;
-			}
-
-
-
-			// 5. Max speed and Average speed, if any. Average speed is NOT overall (effective) speed, but only calculated for "moving" periods.
-			if(speedCount > 0) {
-				if(timeMoving > 0){
-					avgSpeed = (float)totalDistanceMoving / (float)timeMoving * 1000f;
-				} else {
-					avgSpeed = (float)totalSpeedSum / (float)speedCount;
-				}
-			} else {
-				avgSpeed = -1;
-			}
-			return this;
-		}
-		
-	}
-	
-	private static class SplitSegment {
-		TrkSegment  segment;
-		double startCoeff = 0;
-		int startPointInd;
-		double endCoeff = 0;
-		int endPointInd;
-		double metricEnd;
-		double secondaryMetricEnd;
-		
-		public SplitSegment(TrkSegment s) {
-			startPointInd = 0;
-			startCoeff = 0;
-			endPointInd = s.points.size() - 2;
-			endCoeff = 1;
-			this.segment = s;
-		}
-		
-		public SplitSegment(TrkSegment s, int pointInd, double cf) {
-			this.segment = s;
-			this.startPointInd = pointInd;
-			this.startCoeff = cf;
-		}
-		
-		
-		public int getNumberOfPoints() {
-			return endPointInd - startPointInd + 2;
-		}
-		
-		public WptPt get(int j) {
-			final int ind = j + startPointInd;
-			if(j == 0) {
-				if(startCoeff == 0) {
-					return segment.points.get(ind);
-				}
-				return approx(segment.points.get(ind), segment.points.get(ind + 1), startCoeff);
-			}
-			if(j == getNumberOfPoints() - 1) {
-				if(endCoeff == 1) {
-					return segment.points.get(ind);
-				}
-				return approx(segment.points.get(ind - 1), segment.points.get(ind), endCoeff);
-			}
-			return segment.points.get(ind);
-		}
-
-		
-		private WptPt approx(WptPt w1, WptPt w2, double cf) {
-			long time = value(w1.time, w2.time, 0, cf);
-			double speed = value(w1.speed, w2.speed, 0, cf);
-			double ele = value(w1.ele, w2.ele, 0, cf);
-			double hdop = value(w1.hdop, w2.hdop, 0, cf);
-			double lat = value(w1.lat, w2.lat, -360, cf);
-			double lon = value(w1.lon, w2.lon, -360, cf);
-			return new WptPt(lat, lon, time, ele, speed, hdop);
-		}
-		
-		private double value(double vl, double vl2, double none, double cf) {
-			if(vl == none || Double.isNaN(vl)) {
-				return vl2;
-			} else if (vl2 == none || Double.isNaN(vl2)) {
-				return vl;
-			}
-			return vl + cf * (vl2 - vl);
-		}
-
-		private long value(long vl, long vl2, long none, double cf) {
-			if(vl == none) {
-				return vl2;
-			} else if(vl2 == none) {
-				return vl;
-			}
-			return vl + ((long) (cf * (vl2 - vl)));
-		}
 
 	
-		public double setLastPoint(int pointInd, double endCf) {
-			endCoeff = endCf;
-			endPointInd = pointInd;
-			return endCoeff;
-		}
-		
-	}
-	
-	private static SplitMetric getDistanceMetric() {
+	private SplitMetric getDistanceMetric() {
 		return new SplitMetric() {
 			
 			private float[] calculations = new float[1];
@@ -533,7 +93,7 @@ public class GPXUtilities {
 		};
 	}
 	
-	private static SplitMetric getTimeSplit() {
+	private SplitMetric getTimeSplit() {
 		return new SplitMetric() {
 			
 			@Override
@@ -546,13 +106,13 @@ public class GPXUtilities {
 		};
 	}
 	
-	private abstract static class SplitMetric {
+	private abstract class SplitMetric {
 
 		public abstract double metric(WptPt p1, WptPt p2);
 
 	}
 	
-	private static void splitSegment(SplitMetric metric, SplitMetric secondaryMetric,
+	private void splitSegment(SplitMetric metric, SplitMetric secondaryMetric,
 			double metricLimit, List<SplitSegment> splitSegments,
 			TrkSegment segment) {
 		double currentMetricEnd = metricLimit;
@@ -590,8 +150,8 @@ public class GPXUtilities {
 		}
 	}
 
-	private static List<GPXTrackAnalysis> convert(List<SplitSegment> splitSegments) {
-		List<GPXTrackAnalysis> ls = new ArrayList<GPXUtilities.GPXTrackAnalysis>();
+	private List<GPXTrackAnalysis> convert(List<SplitSegment> splitSegments) {
+		List<GPXTrackAnalysis> ls = new ArrayList<GPXTrackAnalysis>();
 		for(SplitSegment s : splitSegments) {
 			GPXTrackAnalysis a = new GPXTrackAnalysis();
 			a.prepareInformation(0, s);
@@ -600,195 +160,14 @@ public class GPXUtilities {
 		return ls;
 	}
 	
-	public static class GPXFile extends GPXExtensions {
-		public String author;
-		public List<Track> tracks = new ArrayList<Track>();
-		public List<WptPt> points = new ArrayList<WptPt>();
-		public List<Route> routes = new ArrayList<Route>();
-		
-		public String warning = null;
-		public String path = "";
-		public boolean showCurrentTrack;
-		public long modifiedTime = 0;
 
-		public boolean isCloudmadeRouteFile() {
-			return "cloudmade".equalsIgnoreCase(author);
-		}
-		
-		
-		public GPXTrackAnalysis getAnalysis(long fileTimestamp) {
-			GPXTrackAnalysis g = new GPXTrackAnalysis();
-			g.wptPoints = points.size();
-			List<SplitSegment> splitSegments = new ArrayList<GPXUtilities.SplitSegment>();
-			for(int i = 0; i< tracks.size() ; i++){
-				Track subtrack = tracks.get(i);
-				for(TrkSegment segment : subtrack.segments){
-					g.totalTracks ++;
-					if(segment.points.size() > 1) {
-						splitSegments.add(new SplitSegment(segment));
-					}
-				}
-			}
-			g.prepareInformation(fileTimestamp, splitSegments.toArray(new SplitSegment[splitSegments.size()]));
-			return g ;
-		}
-
-
-		public boolean hasRtePt() {
-			for(Route r : routes) {
-				if(r.points.size() > 0) {
-					return true;
-				}
-			}
-			return false;
-		}
-		
-		public boolean hasWptPt() {
-			return points.size() > 0;
-		}
-		
-		public boolean hasTrkpt() {
-			for(Track t  : tracks) {
-				for (TrkSegment ts : t.segments) {
-					if (ts.points.size() > 0) {
-						return true;
-					}
-				}
-			}
-			return false;
-		}
-
-		public WptPt addWptPt(double lat, double lon, long time, String description, String name, String category, int color) {
-			double latAdjusted = Double.parseDouble(latLonFormat.format(lat));
-			double lonAdjusted = Double.parseDouble(latLonFormat.format(lon));
-			final WptPt pt = new WptPt(latAdjusted, lonAdjusted, time, Double.NaN, 0, Double.NaN);
-			pt.name = name;
-			pt.category = category;
-			pt.desc = description;
-			if (color != 0) {
-				pt.setColor(color);
-			}
-
-			points.add(pt);
-			modifiedTime = System.currentTimeMillis();
-
-			return pt;
-		}
-
-		public void updateWptPt(WptPt pt, double lat, double lon, long time, String description, String name, String category, int color) {
-			int index = points.indexOf(pt);
-
-			double latAdjusted = Double.parseDouble(latLonFormat.format(lat));
-			double lonAdjusted = Double.parseDouble(latLonFormat.format(lon));
-
-			pt.lat = latAdjusted;
-			pt.lon = lonAdjusted;
-			pt.time = time;
-			pt.desc = description;
-			pt.name = name;
-			pt.category = category;
-			if (color != 0) {
-				pt.setColor(color);
-			}
-
-			if (index != -1) {
-				points.set(index, pt);
-			}
-			modifiedTime = System.currentTimeMillis();
-		}
-
-		public boolean deleteWptPt(WptPt pt) {
-			modifiedTime = System.currentTimeMillis();
-			return points.remove(pt);
-		}
-
-		public List<TrkSegment> processRoutePoints() {
-			List<TrkSegment> tpoints = new ArrayList<TrkSegment>();
-			if (routes.size() > 0) {
-				for (Route r : routes) {
-					TrkSegment sgmt = new TrkSegment();
-					tpoints.add(sgmt);
-					sgmt.points.addAll(r.points);
-				}
-			}
-			return tpoints;
-		}
-
-		public List<TrkSegment> proccessPoints() {
-			List<TrkSegment> tpoints = new ArrayList<TrkSegment>();
-			for (Track t : tracks) {
-				int trackColor = t.getColor(getColor(0));
-				float trackZoom = t.getGpxZoom(getGpxZoom(1.0f));
-				for (TrkSegment ts : t.segments) {
-					if (ts.points.size() > 0) {
-						TrkSegment sgmt = new TrkSegment();
-						sgmt.setCustomZoom(ts.hasCustomZoom()|t.hasCustomZoom());
-						tpoints.add(sgmt);
-						sgmt.points.addAll(ts.points);
-						sgmt.setColor(ts.getColor(trackColor));
-						sgmt.setGpxZoom(ts.getGpxZoom(trackZoom));
-					}
-				}
-			}
-			return tpoints;
-		}
-		
-		public WptPt getLastPoint() {
-			if (tracks.size() > 0) {
-				Track tk = tracks.get(tracks.size() - 1);
-				if (tk.segments.size() > 0) {
-					TrkSegment ts = tk.segments.get(tk.segments.size() - 1);
-					if (ts.points.size() > 0) {
-						return ts.points.get(ts.points.size() - 1);
-					}
-				}
-			}
-			return null;
-		}
-
-		public WptPt findPointToShow() {
-			for (Track t : tracks) {
-				for (TrkSegment s : t.segments) {
-					if (s.points.size() > 0) {
-						return s.points.get(0);
-					}
-				}
-			}
-			for (Route s : routes) {
-				if (s.points.size() > 0) {
-					return s.points.get(0);
-				}
-			}
-			if (points.size() > 0) {
-				return points.get(0);
-			}
-			return null;
-		}
-
-		public boolean isEmpty() {
-			for (Track t : tracks) {
-				if (t.segments != null) {
-					for (TrkSegment s : t.segments) {
-						boolean tracksEmpty = s.points.isEmpty();
-						if (!tracksEmpty) {
-							return false;
-						}
-					}
-				}
-			}
-			return points.isEmpty() && routes.isEmpty();
-		}
-
-
-	}
-
-    public static String asString(GPXFile file, OsmandApplication ctx) {
+    public String asString(GPXFile file, OsmandApplication ctx) {
            final Writer writer = new StringWriter();
-           GPXUtilities.writeGpx(writer, file, ctx);
+           writeGpx(writer, file, ctx);
            return writer.toString();
        }
 
-	public static String writeGpxFile(File fout, GPXFile file, OsmandApplication ctx) {
+	public String writeGpxFile(File fout, GPXFile file, OsmandApplication ctx) {
 		Writer output = null;
 		try {
 			output = new OutputStreamWriter(new FileOutputStream(fout), "UTF-8"); //$NON-NLS-1$
@@ -807,7 +186,7 @@ public class GPXUtilities {
 		}
 	}
 
-	public static String writeGpx(Writer output, GPXFile file, OsmandApplication ctx) {
+	public String writeGpx(Writer output, GPXFile file, OsmandApplication ctx) {
 		try {
 			SimpleDateFormat format = new SimpleDateFormat(GPX_TIME_FORMAT, Locale.US);
 			format.setTimeZone(TimeZone.getTimeZone("UTC"));
@@ -877,7 +256,7 @@ public class GPXUtilities {
 		return null;
 	}
 
-	private static void writeNotNullText(XmlSerializer serializer, String tag, String value) throws IOException {
+	private void writeNotNullText(XmlSerializer serializer, String tag, String value) throws IOException {
 		if (value != null) {
 			serializer.startTag(null, tag);
 			serializer.text(value);
@@ -885,7 +264,7 @@ public class GPXUtilities {
 		}
 	}
 
-	private static void writeExtensions(XmlSerializer serializer, GPXExtensions p) throws IOException {
+	private void writeExtensions(XmlSerializer serializer, GPXExtensions p) throws IOException {
 		if (!p.getExtensionsToRead().isEmpty()) {
 			serializer.startTag(null, "extensions");
 			for (Map.Entry<String, String> s : p.getExtensionsToRead().entrySet()) {
@@ -895,7 +274,10 @@ public class GPXUtilities {
 		}
 	}
 
-	private static void writeWpt(SimpleDateFormat format, XmlSerializer serializer, WptPt p) throws IOException {
+	private void writeWpt(SimpleDateFormat format, XmlSerializer serializer, WptPt p) throws IOException {
+
+		final NumberFormat latLonFormat = new DecimalFormat("0.00#####", new DecimalFormatSymbols(new Locale("EN", "US")));
+
 		serializer.attribute(null, "lat", latLonFormat.format(p.lat)); //$NON-NLS-1$ //$NON-NLS-2$
 		serializer.attribute(null, "lon", latLonFormat.format(p.lon)); //$NON-NLS-1$ //$NON-NLS-2$
 
@@ -922,7 +304,7 @@ public class GPXUtilities {
 		writeExtensions(serializer, p);
 	}
 
-	public static class GPXFileResult {
+	public class GPXFileResult {
 		public ArrayList<List<Location>> locations = new ArrayList<List<Location>>();
 		public ArrayList<WptPt> wayPoints = new ArrayList<WptPt>();
 		// special case for cloudmate gpx : they discourage common schema
@@ -943,7 +325,7 @@ public class GPXUtilities {
 		}
 	}
 
-	private static String readText(XmlPullParser parser, String key) throws XmlPullParserException, IOException {
+	private String readText(XmlPullParser parser, String key) throws XmlPullParserException, IOException {
 		int tok;
 		String text = null;
 		while ((tok = parser.next()) != XmlPullParser.END_DOCUMENT) {
@@ -961,7 +343,7 @@ public class GPXUtilities {
 		return text;
 	}
 
-	public static GPXFile loadGPXFile(Context ctx, File f) {
+	public GPXFile loadGPXFile(Context ctx, File f) {
 		FileInputStream fis = null;
 		try {
 			fis = new FileInputStream(f);
@@ -988,6 +370,7 @@ public class GPXUtilities {
 		}
 	}
 
+	/* original version
 	public static GPXFile loadGPXFile(Context ctx, InputStream f) {
 		GPXFile res = new GPXFile();
 		SimpleDateFormat format = new SimpleDateFormat(GPX_TIME_FORMAT, Locale.US);
@@ -1162,8 +545,176 @@ public class GPXUtilities {
 
 		return res;
 	}
+	*/
 
-	private static Reader getUTF8Reader(InputStream f) throws IOException {
+	public GPXFile loadGPXFile(Context ctx, InputStream f) {
+		GPXFile res = new GPXFile();
+		SimpleDateFormat format = new SimpleDateFormat(GPX_TIME_FORMAT, Locale.US);
+		format.setTimeZone(TimeZone.getTimeZone("UTC"));
+		try {
+
+			XmlPullParser parser = PlatformUtil.newXMLPullParser();
+			parser.setInput(getUTF8Reader(f)); //$NON-NLS-1$
+			Stack<GPXExtensions> parserState = new Stack<GPXExtensions>();
+			boolean extensionReadMode = false;
+			parserState.push(res);
+			int tok;
+			while ((tok = parser.next()) != XmlPullParser.END_DOCUMENT) {
+				if (tok == XmlPullParser.START_TAG) {
+					Object parse = parserState.peek();
+					String tag = parser.getName();
+					if (extensionReadMode && parse instanceof GPXExtensions) {
+						String value = readText(parser, tag);
+						if (value != null) {
+							((GPXExtensions) parse).getExtensionsToWrite().put(tag, value);
+							if (tag.equals("speed") && parse instanceof WptPt) {
+								try {
+									((WptPt) parse).speed = Float.parseFloat(value);
+								} catch (NumberFormatException e) {
+								}
+							}
+						}
+
+					} else if (parse instanceof GPXExtensions && tag.equals("extensions")) {
+						extensionReadMode = true;
+					} else {
+
+						if (parse instanceof TrkSegment) {						// 1st for speed
+							if (tag.equals("trkpt")) {
+								WptPt wptPt = parseWptAttributes(parser);
+								((TrkSegment) parse).points.add(wptPt);
+								parserState.push(wptPt);
+							}
+						} else if (parse instanceof WptPt) {
+							if (tag.equals("name")) {
+								((WptPt) parse).name = readText(parser, tag);
+							} else if (tag.equals("desc")) {
+								((WptPt) parse).desc = readText(parser, tag);
+							} else if (tag.equals("link")) {
+								((WptPt) parse).link = parser.getAttributeValue("", "href");
+							} else if (tag.equals("category")) {
+								((WptPt) parse).category = readText(parser, tag);
+							} else if (tag.equals("type")) {
+								if(((WptPt) parse).category == null) {
+									((WptPt) parse).category = readText(parser, tag);
+								}
+							} else if (tag.equals("ele")) {
+								String text = readText(parser, tag);
+								if (text != null) {
+									try {
+										((WptPt) parse).ele = Float.parseFloat(text);
+									} catch (NumberFormatException e) {
+									}
+								}
+							} else if (tag.equals("hdop")) {
+								String text = readText(parser, tag);
+								if (text != null) {
+									try {
+										((WptPt) parse).hdop = Float.parseFloat(text);
+									} catch (NumberFormatException e) {
+									}
+								}
+							} else if (tag.equals("time")) {
+								String text = readText(parser, tag);
+								if (text != null) {
+									try {
+										((WptPt) parse).time = format.parse(text).getTime();
+									} catch (ParseException e) {
+									}
+								}
+							}
+						} else if (parse instanceof GPXFile) {
+							if (tag.equals("wpt")) {							// moved 1st for speed
+								WptPt wptPt = parseWptAttributes(parser);
+								((GPXFile) parse).points.add(wptPt);
+								parserState.push(wptPt);
+							} else if (tag.equals("gpx")) {
+								((GPXFile) parse).author = parser.getAttributeValue("", "creator");
+							} else if (tag.equals("trk")) {
+								Track track = new Track();
+								((GPXFile) parse).tracks.add(track);
+								parserState.push(track);
+							} else if (tag.equals("rte")) {
+								Route route = new Route();
+								((GPXFile) parse).routes.add(route);
+								parserState.push(route);
+							}
+						} else if (parse instanceof Route) {
+							if (tag.equals("name")) {
+								((Route) parse).name = readText(parser, tag);
+							} else if (tag.equals("desc")) {
+								((Route) parse).desc = readText(parser, tag);
+							} else if (tag.equals("rtept")) {
+								WptPt wptPt = parseWptAttributes(parser);
+								((Route) parse).points.add(wptPt);
+								parserState.push(wptPt);
+							}
+						} else if (parse instanceof Track) {
+							if (tag.equals("name")) {
+								((Track) parse).name = readText(parser, tag);
+							} else if (tag.equals("desc")) {
+								((Track) parse).desc = readText(parser, tag);
+							} else if (tag.equals("trkseg")) {
+								TrkSegment trkSeg = new TrkSegment();
+								((Track) parse).segments.add(trkSeg);
+								parserState.push(trkSeg);
+							}
+						}
+					}
+
+				} else if (tok == XmlPullParser.END_TAG) {
+					String tag = parser.getName();
+					if (tag.equals("trkpt")) {					// moved first for speed!
+						Object pop = parserState.pop();
+						assert pop instanceof WptPt;
+					} else if (tag.equals("wpt")) {
+						Object pop = parserState.pop();
+						assert pop instanceof WptPt;
+					} else if (tag.equals("rtept")) {
+						Object pop = parserState.pop();
+						assert pop instanceof WptPt;
+					} else if (tag.equals("trk")) {
+						Object pop = parserState.pop();
+						assert pop instanceof Track;
+					} else if (tag.equals("rte")) {
+						Object pop = parserState.pop();
+						assert pop instanceof Route;
+					} else if (tag.equals("trkseg")) {
+						Object pop = parserState.pop();
+						assert pop instanceof TrkSegment;
+					} else if (tag.equals("extensions")) {
+						if (parserState.peek() instanceof GPXExtensions)
+							extensionReadMode = false;
+					}
+				}
+			}
+//			if (convertCloudmadeSource && res.isCloudmadeRouteFile()) {
+//				Track tk = new Track();
+//				res.tracks.add(tk);
+//				TrkSegment segment = new TrkSegment();
+//				tk.segments.add(segment);
+//
+//				for (WptPt wp : res.points) {
+//					segment.points.add(wp);
+//				}
+//				res.points.clear();
+//			}
+		} catch (RuntimeException e) {
+			log.error("Error reading gpx", e); //$NON-NLS-1$
+			res.warning = ctx.getString(R.string.error_reading_gpx) + " " + e.getMessage();
+		} catch (XmlPullParserException e) {
+			log.error("Error reading gpx", e); //$NON-NLS-1$
+			res.warning = ctx.getString(R.string.error_reading_gpx) + " " + e.getMessage();
+		} catch (IOException e) {
+			log.error("Error reading gpx", e); //$NON-NLS-1$
+			res.warning = ctx.getString(R.string.error_reading_gpx) + " " + e.getMessage();
+		}
+
+		return res;
+	}
+
+
+	private Reader getUTF8Reader(InputStream f) throws IOException {
 		BufferedInputStream bis = new BufferedInputStream(f);
 		assert bis.markSupported();
 		bis.mark(3);
@@ -1179,7 +730,7 @@ public class GPXUtilities {
 		return new InputStreamReader(bis, "UTF-8");
 	}
 
-	private static WptPt parseWptAttributes(XmlPullParser parser) {
+	private WptPt parseWptAttributes(XmlPullParser parser) {
 		WptPt wpt = new WptPt();
 		try {
 			wpt.lat = Double.parseDouble(parser.getAttributeValue("", "lat")); //$NON-NLS-1$ //$NON-NLS-2$
@@ -1189,7 +740,7 @@ public class GPXUtilities {
 		return wpt;
 	}
 
-	public static void mergeGPXFileInto(GPXFile to, GPXFile from) {
+	public void mergeGPXFileInto(GPXFile to, GPXFile from) {
 		if (from == null) {
 			return;
 		}
@@ -1212,3 +763,5 @@ public class GPXUtilities {
 	}
 
 }
+
+
