+++ OsmAnd/src/net/osmand/plus/RenderType.java
@@ -1,0 +1,11 @@
+package net.osmand.plus;
+
+public enum RenderType {
+	FORCE_REDRAW,
+	ORIGINAL,
+	OUTLINE,
+	RESAMPLE,
+	M1000,
+	M100,
+	CONVEYOR,
+}
