+++ OsmAnd/src/net/osmand/plus/Route.java
@@ -1,0 +1,10 @@
+package net.osmand.plus;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Route extends GPXExtensions {
+	public String name = null;
+	public String desc = null;
+	public List<WptPt> points = new ArrayList<WptPt>();
+}
