+++ OsmAnd/src/net/osmand/plus/Track.java
@@ -1,0 +1,10 @@
+package net.osmand.plus;
+
+import java.util.ArrayList;
+import java.util.List;
+
+public class Track extends GPXExtensions {
+    public String name = null;
+    public String desc = null;
+    public List<TrkSegment> segments = new ArrayList<TrkSegment>();
+}
