+++ plugins/Osmand-SRTMPlugin/build.gradle
@@ -64,4 +64,5 @@ repositories {
 }
 
 dependencies {
+	compile "com.android.support:appcompat-v7:23.0.+"
 }
